Added `de/lists` module
