+++ crates/bevy_reflect/src/serde/de/deserializer.rs
@@ -1,8 +1,9 @@
 use crate::serde::de::arrays::ArrayVisitor;
+use crate::serde::de::lists::ListVisitor;
 use crate::serde::de::structs::StructVisitor;
 use crate::serde::de::tuple_structs::TupleStructVisitor;
 use crate::serde::de::tuples::TupleVisitor;
-use crate::serde::de::{EnumVisitor, ListVisitor, MapVisitor, OptionVisitor, SetVisitor};
+use crate::serde::de::{EnumVisitor, MapVisitor, OptionVisitor, SetVisitor};
 use crate::serde::TypeRegistrationDeserializer;
 use crate::{PartialReflect, ReflectDeserialize, TypeInfo, TypeRegistration, TypeRegistry};
 use core::fmt::Formatter;
@@ -269,10 +270,8 @@ impl<'a, 'de> DeserializeSeed<'de> for TypedReflectDeserializer<'a> {
                 Ok(Box::new(dynamic_tuple_struct))
             }
             TypeInfo::List(list_info) => {
-                let mut dynamic_list = deserializer.deserialize_seq(ListVisitor {
-                    list_info,
-                    registry: self.registry,
-                })?;
+                let mut dynamic_list =
+                    deserializer.deserialize_seq(ListVisitor::new(list_info, self.registry))?;
                 dynamic_list.set_represented_type(Some(self.registration.type_info()));
                 Ok(Box::new(dynamic_list))
             }
